finish project

+++ modules/Eagle/Car/src/Http/Controllers/CarController.php
@@ -9,6 +9,7 @@ use Eagle\Car\Repositories\CarRepo;
 use Eagle\Parking\Repositories\ParkingRepo;
 use Eagle\Payment\Gateways\Gateway;
 use Eagle\Payment\Services\PaymentServices;
+use Eagle\RolePermission\Models\Permission;
 
 class CarController extends Controller
 {
@@ -22,52 +23,75 @@ class CarController extends Controller
 
     public function index()
     {
-        $cars = $this->carRepo->paginate();
+        if (auth()->user()->hasPermissionTo(Permission::SUPER_ADMIN)) {
+            $cars = $this->carRepo->all();
+        } else {
+            $cars = $this->carRepo->findByUserId(auth()->id());
+        }
         return view("Car::index", compact("cars"));
     }
 
     public function create()
     {
+        if (auth()->user()->cars->count()) {
+            newFeedback("شما یک خودرو دارید", "حداکثر خودرو برای شخص یک خودرو می باشد!", "red");
+            return back();
+        }
+
         return view("Car::create");
     }
 
     public function store(StoreCarRequest $request)
     {
+        if (auth()->user()->cars->count()) {
+            newFeedback("شما یک خودرو دارید", "حداکثر خودرو برای شخص یک خودرو می باشد!", "red");
+            return back();
+        }
         $this->carRepo->store($request->all());
         newFeedback("عملیات موفقیت آمیز", "عملیات با موفقیت انجام شد", "green");
-        return redirect()->route("car.index");
+        return redirect("/car");
 
     }
 
     public function edit($id)
     {
         $car = $this->carRepo->findOrFailById($id);
+        $this->authorize("functions", $car);
         return view("Car::edit", compact("car"));
     }
 
     public function update($id, UpdateCarRequest $request)
     {
+        $car = $this->carRepo->findOrFailById($id);
+        $this->authorize("functions", $car);
         $this->carRepo->update($id, $request->all());
         newFeedback("عملیات موفقیت آمیز", "ویرایش با موفقیت انجام شد", "green");
 
-        return redirect()->route("car.index");
+        return redirect("/car");
     }
 
     public function delete($id)
     {
+        $car = $this->carRepo->findOrFailById($id);
+        $this->authorize("functions", $car);
+
         $this->carRepo->delete($id);
         newFeedback("عملیات موفقیت آمیز", "حذف با موفقیت انجام شد", "red");
 
-        return route("car.index");
+        return redirect("/car");
     }
 
-    public function buy($parking_id,ParkingRepo $parkingRepo)
+    public function buy($parking_id, ParkingRepo $parkingRepo)
     {
         $parking = $parkingRepo->findOrFailById($parking_id);
+        $rezervCount = $parking->cars->count();
 
-//        todo feedback
+        if ($parking->capacity <= $rezervCount){
+            newFeedback("ظرفیت پر","ظرفیت این پارکینگ پر است!","red");
+            return back();
+        }
 
-        $payment = PaymentServices::generate($parking->price, $parking, auth()->user(),$parking->user_id);
+        $payment = PaymentServices::generate($parking->price, $parking, auth()->user(), $parking->user_id);
         resolve(Gateway::class)->redirect();
     }
 }
