+++ modules/Eagle/Parking/src/Http/Controllers/ParkingController.php
@@ -3,11 +3,13 @@
 namespace Eagle\Parking\Http\Controllers;
 
 use App\Http\Controllers\Controller;
+use Eagle\Car\Models\Car;
 use Eagle\Parking\Http\Requests\StoreParkingRequest;
 use Eagle\Parking\Http\Requests\UpdateParkingRequest;
 use Eagle\Parking\Models\Parking;
 use Eagle\Parking\Repositories\ParkingRepo;
 use Eagle\Payment\Models\Payment;
+use Illuminate\Support\Facades\DB;
 
 class ParkingController extends Controller
 {
@@ -20,19 +22,21 @@ class ParkingController extends Controller
 
     public function index()
     {
-
+        $this->authorize("manage", Parking::class);
         $parkings = $this->parkingRepo->paginate();
         return view("Parking::index", compact("parkings"));
     }
 
     public function create()
     {
+        $this->authorize("manage", Parking::class);
         return view("Parking::create");
 
     }
 
     public function store(StoreParkingRequest $request)
     {
+        $this->authorize("manage", Parking::class);
         $this->parkingRepo->store($request->all());
         newFeedback("عملیات موفقیت آمیز", "عملیات با موفقیت انجام شد", "green");
 
@@ -41,14 +45,16 @@ class ParkingController extends Controller
 
     public function edit($id)
     {
+        $this->authorize("manage", Parking::class);
         $parking = $this->parkingRepo->findOrFailById($id);
         return view("Parking::edit", compact("parking"));
 
     }
 
-    public function update($id,UpdateParkingRequest $request)
+    public function update($id, UpdateParkingRequest $request)
     {
-        $this->parkingRepo->update($request->all(),$id);
+        $this->authorize("manage", Parking::class);
+        $this->parkingRepo->update($request->all(), $id);
         newFeedback("عملیات موفقیت آمیز", "ویرایش با موفقیت انجام شد", "green");
 
         return redirect()->route("parking.index");
@@ -56,6 +62,7 @@ class ParkingController extends Controller
 
     public function delete($id)
     {
+        $this->authorize("manage", Parking::class);
         $this->parkingRepo->delete($id);
         newFeedback("عملیات موفقیت آمیز", "حذف با موفقیت انجام شد", "red");
 
@@ -65,7 +72,8 @@ class ParkingController extends Controller
     public function detail($id)
     {
         $parking = Parking::query()->findOrFail($id);
-        return view("Parking::detail",compact("parking"));
+        $rezervCount = $parking->cars()->count();
+        return view("Parking::detail", compact("parking","rezervCount"));
     }
